common: audit follow-up — bug fixes, perf, C89 cleanup (round 2)
Second pass on top of fe0216e, working through the follow-up checklist
from the initial audit.

Bug fixes:
- net_dgrm: cast (int)sizeof(int) in the five `if (len/ret < sizeof(int))`
  guards so a negative read result (-1 from driver->Read on error)
  correctly breaks the loop instead of being unsigned-promoted to
  SIZE_MAX, which would let the code continue with cursize = -1.
- alias_model: drop the (uintptr_t)((daliasskingroup_t*)skinnum)
  double-cast. skinnum is just a counter; type it int. firstframe in
  maliasskindesc_t is just an array index — also int. Saves 4 bytes per
  skin descriptor on 64-bit and removes the laundering pointer cast.

Performance:
- r_surf: R_AddDynamicLights and R_AddDynamicLightsRGB iterate set bits
  via compat_ctz() over each word of dlightbits[], skipping empty slots
  entirely instead of testing all MAX_DLIGHTS bits per surface. Robust
  to MAX_DLIGHTS growing past 32 (loops over each word).
- mathlib, r_bsp, r_part, r_sprite, sv_user, sv_move, sv_phys, view,
  pr_cmds, cl_tent, model, r_main, drawhulls, d_sprite, d_polyse,
  r_draw: convert sin/cos/tan/atan2/sqrt/fabs/floor/ceil to their
  single-precision (f-suffixed) variants in float-only contexts.
  Removes a series of float→double promote / double→float narrow
  pairs around every call. Particularly relevant on 32-bit ARM where
  the FPU is float-only and emulating double is expensive. The handful
  of legitimately double-precision sites (FloorDivMod, the explicit
  (double) cast in R_ViewChanged) are intentionally left alone.

Cleanup:
- zone: pad hunk_t to 16 bytes (added int pad[2]) so the (h+1)
  pointer returned by Hunk_Alloc actually honors the 16-byte alignment
  documented in zone.h. Previously sizeof(hunk_t) was 8 so returned
  pointers were only 8-byte aligned; fine for the C scalar code, but
  wrong for any SIMD-aligned consumer.
- quakedef: replace the narrow _WIN32-only `#define __func__
  __FUNCTION__` shim with logic that activates only on truly strict-C89
  compilers without either __func__ or __FUNCTION__. GCC, Clang, MSVC
  and ICC accept __func__ as an extension regardless of -std=, so the
  fallback no longer fires for them and the warning count under
  -std=gnu89 -pedantic drops sharply.
- net_win, drawhulls: replace C99 designated initializers (.name = ...)
  with positional initializers and field-name comments, restoring
  C89 compliance. drawhulls is GLquake-only; the cvar_t init handles
  both NQ_HACK and QW_HACK due to a conditional struct member.

Verified: all 88 common/*.c files compile under
gcc -std=gnu89 -pedantic -Wdeclaration-after-statement -Wno-comment,
zero errors, only -Wformat-truncation advisories on snprintf calls
remain (defined behavior, cosmetic).

+++ common/cl_tent.c
@@ -339,12 +339,12 @@ void CL_UpdateTEnts(void)
       }
       else
       {
-         yaw = (int)(atan2(dist[1], dist[0]) * 180 / M_PI);
+         yaw = (int)(atan2f(dist[1], dist[0]) * 180 / M_PI);
          if (yaw < 0)
             yaw += 360;
 
-         forward = sqrt(dist[0] * dist[0] + dist[1] * dist[1]);
-         pitch = (int)(atan2(dist[2], forward) * 180 / M_PI);
+         forward = sqrtf(dist[0] * dist[0] + dist[1] * dist[1]);
+         pitch = (int)(atan2f(dist[2], forward) * 180 / M_PI);
          if (pitch < 0)
             pitch += 360;
       }
