+++ common/d_polyse.c
@@ -589,8 +589,8 @@ void D_PolysetCalcGradients(int skinwidth)
    /* Underflow is very visible, overflow is very unlikely, because of ambient lighting */
    t0 = r_p0[4] - r_p2[4];
    t1 = r_p1[4] - r_p2[4];
-   r_lstepx  = (int) ceil (t1 * p01_minus_p21 - t0 * p11_minus_p21);
-   r_lstepy  = (int) ceil (t1 * p00_minus_p20 - t0 * p10_minus_p20);
+   r_lstepx  = (int) ceilf (t1 * p01_minus_p21 - t0 * p11_minus_p21);
+   r_lstepy  = (int) ceilf (t1 * p00_minus_p20 - t0 * p10_minus_p20);
 
    t0 = r_p0[5] - r_p2[5];
    t1 = r_p1[5] - r_p2[5];
