+++ common/model.c
@@ -1019,8 +1019,8 @@ CalcSurfaceExtents(msurface_t *s)
     }
 
     for (i = 0; i < 2; i++) {
-	bmins[i] = floor(mins[i] / 16);
-	bmaxs[i] = ceil(maxs[i] / 16);
+	bmins[i] = floorf(mins[i] / 16);
+	bmaxs[i] = ceilf(maxs[i] / 16);
 
 	s->texturemins[i] = bmins[i] * 16;
 	s->extents[i] = (bmaxs[i] - bmins[i]) * 16;
@@ -1807,7 +1807,7 @@ static float RadiusFromBounds(vec3_t mins, vec3_t maxs)
    vec3_t corner;
 
    for (i = 0; i < 3; i++)
-      corner[i] = qmax(fabs(mins[i]), fabs(maxs[i]));
+      corner[i] = qmax(fabsf(mins[i]), fabsf(maxs[i]));
 
    return Length(corner);
 }
