+++ common/r_draw.c
@@ -100,7 +100,7 @@ void R_EmitEdge(mvertex_t *pv0, mvertex_t *pv1)
       if (v0 > r_refdef.fvrectbottom_adj)
          v0 = r_refdef.fvrectbottom_adj;
 
-      ceilv0 = (int)ceil(v0);
+      ceilv0 = (int)ceilf(v0);
    }
 
    world = &pv1->position[0];
@@ -140,7 +140,7 @@ void R_EmitEdge(mvertex_t *pv0, mvertex_t *pv1)
 
    r_emitted = 1;
 
-   r_ceilv1 = (int)ceil(r_v1);
+   r_ceilv1 = (int)ceilf(r_v1);
 
    /* create the edge */
    if (ceilv0 == r_ceilv1) {
