+++ common/r_sprite.c
@@ -330,8 +330,8 @@ void R_DrawSprite(const entity_t *e)
       /* that plane around the center according to the sprite entity's roll */
       /* angle. So vpn stays the same, but vright and vup rotate */
       angle = e->angles[ROLL] * (M_PI * 2 / 360);
-      sr = sin(angle);
-      cr = cos(angle);
+      sr = sinf(angle);
+      cr = cosf(angle);
 
       for (i = 0; i < 3; i++) {
          r_spritedesc.vpn[i] = vpn[i];
