+++ common/sv_phys.c
@@ -839,8 +839,8 @@ SV_TryUnstick(edict_t *ent, vec3_t oldvel)
       clip               = SV_FlyMove(ent, 0.1, &steptrace);
 
       if (
-            fabs(oldorg[1] - ent->v.origin[1]) > 4 ||
-            fabs(oldorg[0] - ent->v.origin[0]) > 4)
+            fabsf(oldorg[1] - ent->v.origin[1]) > 4 ||
+            fabsf(oldorg[0] - ent->v.origin[0]) > 4)
       {
          /* Con_DPrintf ("unstuck!\n"); */
          return clip;
@@ -921,8 +921,8 @@ SV_WalkMove(edict_t *ent)
    if (clip)
    {
       if (
-               fabs(oldorg[1] - ent->v.origin[1]) < 0.03125 
-            && fabs(oldorg[0] - ent->v.origin[0]) < 0.03125)
+               fabsf(oldorg[1] - ent->v.origin[1]) < 0.03125
+            && fabsf(oldorg[0] - ent->v.origin[0]) < 0.03125)
       {
          /* stepping up didn't make any progress */
          clip = SV_TryUnstick(ent, oldvel);
